prevent unregistered events from triggering again
Noticed that sometimes events (RecordingStarted, I'm looking at you) fire multiple times unexpectedly, and if they do it quickly enough, a "one-time listener" won't be able to be unregistered quickly enough. This solution is improvement in trying to beat the race condition, but because the handler is scheduled on the task runner, it might still miss - implementors of one-time listeners should protect themselves better as well.