destroy bridge should retry on any error other than destroy bridge.
The reason is that we saw cases where destroy bridge would throw "channel not found" (which is not a valid error for this operation) - maybe because it tried to kick all the channels first and one of the channels was disconnected in the middle of a race condition. So just retry.