retry ARI operations when they fail with 500 code
We're assuming Asterisk glitches from time to time (we get 500 errors with no crash or any error report in the Asterisk logs), so just retry - no need to map as we don't expect useful info in internal error (but log it anyway).

Also, try not to lose the caller's stack when retrying.